refactor: top activity (collection, buyer, seller)
