+++ components/Activity/CollectionTransactionDetail.js
@@ -1,30 +1,13 @@
 import axios from 'axios'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 import { useEffect, useState } from 'react'
 import Scrollbars from 'react-custom-scrollbars'
-
-import Card from '../Card/Card'
 
 import { parseImgUrl } from 'utils/common'
 import { useIntl } from 'hooks/useIntl'
 import { formatNearAmount } from 'near-api-js/lib/utils/format'
-import cachios from 'cachios'
 import TokenDetailModal from 'components/Token/TokenDetailModal'
-
-const renderThumb = ({ style, ...props }) => {
-	return (
-		<div
-			{...props}
-			style={{
-				...style,
-				cursor: 'pointer',
-				borderRadius: 'inherit',
-				backgroundColor: 'rgba(255, 255, 255, 0.2)',
-			}}
-		/>
-	)
-}
+import TopTransactionCard, { renderThumb } from './TopTransactionCard'
 
 const CollectionTransactionList = ({ data }) => {
 	const [localToken, setLocalToken] = useState(null)
@@ -33,7 +16,7 @@ const CollectionTransactionList = ({ data }) => {
 		<>
 			<TokenDetailModal tokens={[localToken]} />
 			{data.map((user, idx) => (
-				<UserTransactionDetail
+				<CollectionTransactionDetail
 					data={user}
 					key={user.account_id}
 					idx={idx}
@@ -44,7 +27,7 @@ const CollectionTransactionList = ({ data }) => {
 	)
 }
 
-const UserTransactionDetail = ({ data, idx, setLocalToken }) => {
+const CollectionTransactionDetail = ({ data, idx, setLocalToken }) => {
 	const { localeLn } = useIntl()
 
 	const [colDetail, setColDetail] = useState({})
@@ -84,7 +67,7 @@ const UserTransactionDetail = ({ data, idx, setLocalToken }) => {
 							</Link>
 						)}
 						<p className="text-base text-gray-400">
-							Total sales: {formatNearAmount(data.total_sum)} Ⓝ
+							Total sales: {formatNearAmount(data.total_sum, 2)} Ⓝ
 						</p>
 						<p className="text-base text-gray-400">
 							{localeLn('Card')} sold: {data.contract_token_ids.length}
@@ -103,7 +86,7 @@ const UserTransactionDetail = ({ data, idx, setLocalToken }) => {
 				>
 					<div className="flex py-2">
 						{data.contract_token_ids.slice(0, 5).map((contract_token_id, idx) => (
-							<UserTransactionCard
+							<TopTransactionCard
 								key={idx}
 								contract_token_id={contract_token_id}
 								setLocalToken={setLocalToken}
@@ -111,74 +94,6 @@ const UserTransactionDetail = ({ data, idx, setLocalToken }) => {
 						))}
 					</div>
 				</Scrollbars>
-			</div>
-		</div>
-	)
-}
-
-const UserTransactionCard = ({ contract_token_id, setLocalToken }) => {
-	const router = useRouter()
-	const [token, setToken] = useState(null)
-
-	const [contractId, tokenId] = contract_token_id.split('::')
-
-	useEffect(() => {
-		fetchData()
-	}, [])
-
-	const fetchData = async () => {
-		const params = {
-			contract_id: contractId,
-			token_id: tokenId,
-		}
-		const resp = await cachios.get(`${process.env.V2_API_URL}/token`, {
-			params: params,
-			ttl: 60,
-		})
-		setToken(resp.data.data.results[0])
-	}
-
-	if (!token) {
-		return null
-	}
-
-	return (
-		<div
-			id={contract_token_id}
-			className="w-1/3 md:w-1/5 px-2 inline-block whitespace-normal overflow-visible flex-shrink-0"
-		>
-			<div className="w-full m-auto" onClick={() => setLocalToken(token)}>
-				<Card
-					imgUrl={parseImgUrl(token.metadata.media, null, {
-						width: `300`,
-						useOriginal: process.env.APP_ENV === 'production' ? false : true,
-						isMediaCdn: token.isMediaCdn,
-					})}
-					onClick={() => {
-						router.push(
-							{
-								pathname: router.pathname,
-								query: {
-									...router.query,
-									tokenId: token.token_id,
-									contractId: token.contract_id,
-								},
-							},
-							`/token/${token.contract_id}::${token.token_series_id}/${token.token_id}`,
-							{
-								shallow: true,
-								scroll: false,
-							}
-						)
-					}}
-					imgBlur={token.metadata.blurhash}
-					token={{
-						title: token.metadata.title,
-						collection: token.metadata.collection || token.contract_id,
-						copies: token.metadata.copies,
-						creatorId: token.metadata.creator_id || token.contract_id,
-					}}
-				/>
 			</div>
 		</div>
 	)
