+++ components/Activity/UserTransactionDetail.js
@@ -1,32 +1,16 @@
 import axios from 'axios'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 import { useEffect, useState } from 'react'
 import Scrollbars from 'react-custom-scrollbars'
 
-import Card from '../Card/Card'
 import LinkToProfile from '../LinkToProfile'
 
 import { parseImgUrl } from 'utils/common'
 import InfiniteScroll from 'react-infinite-scroll-component'
 import { useIntl } from 'hooks/useIntl'
 import { formatNearAmount } from 'near-api-js/lib/utils/format'
-import cachios from 'cachios'
 import TokenDetailModal from 'components/Token/TokenDetailModal'
-
-const renderThumb = ({ style, ...props }) => {
-	return (
-		<div
-			{...props}
-			style={{
-				...style,
-				cursor: 'pointer',
-				borderRadius: 'inherit',
-				backgroundColor: 'rgba(255, 255, 255, 0.2)',
-			}}
-		/>
-	)
-}
+import TopTransactionCard, { renderThumb } from './TopTransactionCard'
 
 const UserTransactionList = ({ usersData, fetchData, hasMore, type }) => {
 	const [localToken, setLocalToken] = useState(null)
@@ -85,7 +69,8 @@ const UserTransactionDetail = ({ data, idx, type = 'buyer', setLocalToken }) => 
 							/>
 						)}
 						<p className="text-base text-gray-400">
-							Total {type !== 'buyer' ? 'sales' : 'purchase'}: {formatNearAmount(data.total_sum)} Ⓝ
+							Total {type !== 'buyer' ? 'sales' : 'purchase'}: {formatNearAmount(data.total_sum, 2)}{' '}
+							Ⓝ
 						</p>
 						<p className="text-base text-gray-400">
 							{localeLn('Card')} {type !== 'buyer' ? 'sold' : 'bought'}:{' '}
@@ -105,7 +90,7 @@ const UserTransactionDetail = ({ data, idx, type = 'buyer', setLocalToken }) => 
 				>
 					<div className="flex py-2">
 						{data.contract_token_ids.slice(0, 5).map((contract_token_id, idx) => (
-							<UserTransactionCard
+							<TopTransactionCard
 								key={idx}
 								contract_token_id={contract_token_id}
 								setLocalToken={setLocalToken}
@@ -113,74 +98,6 @@ const UserTransactionDetail = ({ data, idx, type = 'buyer', setLocalToken }) => 
 						))}
 					</div>
 				</Scrollbars>
-			</div>
-		</div>
-	)
-}
-
-const UserTransactionCard = ({ contract_token_id, setLocalToken }) => {
-	const router = useRouter()
-	const [token, setToken] = useState(null)
-
-	const [contractId, tokenId] = contract_token_id.split('::')
-
-	useEffect(() => {
-		fetchData()
-	}, [])
-
-	const fetchData = async () => {
-		const params = {
-			contract_id: contractId,
-			token_id: tokenId,
-		}
-		const resp = await cachios.get(`${process.env.V2_API_URL}/token`, {
-			params: params,
-			ttl: 60,
-		})
-		setToken(resp.data.data.results[0])
-	}
-
-	if (!token) {
-		return null
-	}
-
-	return (
-		<div
-			id={contract_token_id}
-			className="w-1/3 md:w-1/5 px-2 inline-block whitespace-normal overflow-visible flex-shrink-0"
-		>
-			<div className="w-full m-auto" onClick={() => setLocalToken(token)}>
-				<Card
-					imgUrl={parseImgUrl(token.metadata.media, null, {
-						width: `600`,
-						useOriginal: process.env.APP_ENV === 'production' ? false : true,
-						isMediaCdn: token.isMediaCdn,
-					})}
-					onClick={() => {
-						router.push(
-							{
-								pathname: router.pathname,
-								query: {
-									...router.query,
-									tokenId: token.token_id,
-									contractId: token.contract_id,
-								},
-							},
-							`/token/${token.contract_id}::${token.token_series_id}/${token.token_id}`,
-							{
-								shallow: true,
-								scroll: false,
-							}
-						)
-					}}
-					imgBlur={token.metadata.blurhash}
-					token={{
-						title: token.metadata.title,
-						collection: token.metadata.collection || token.contract_id,
-						copies: token.metadata.copies,
-						creatorId: token.metadata.creator_id || token.contract_id,
-					}}
-				/>
 			</div>
 		</div>
 	)
